RF_15.001.1 | User > CRUD user
Increase stability
Remove LoginTest to outer class
